loosen version constraints on test-kitchen
Current version is 2.3.3 and the transport still works.

I also removed `version.rb` file and hardcoded the version as a string 
in
gemspec to avoid loading issues.

`Kitchen::Transport::Rsync` is a class, but the version file defined it
as a module and kitchen would fail loading the plugin:

```
TypeError: Rsync is not a class
  
/home/rob/projects/kitchen-transport-rsync/lib/kitchen/transport/rsync.rb:23:in 
`<module:Transport>'
  
/home/rob/projects/kitchen-transport-rsync/lib/kitchen/transport/rsync.rb:22:in 
`<module:Kitchen>'
  
/home/rob/projects/kitchen-transport-rsync/lib/kitchen/transport/rsync.rb:21:in 
`<top (required)>'
```

There are no references to `VERSION` constant outside of gemspec file.